+++ README.md
@@ -1,15 +1,13 @@
-# StructuredLogger
+# Logr
 
-Welcome to your new gem! In this directory, you'll find the files you need to be able to package up your Ruby library into a gem. Put your Ruby code in the file `lib/structured_logger`. To experiment with that code, run `bin/console` for an interactive prompt.
-
-TODO: Delete this and the text above, and describe your gem
+Structured logging with metrics and events.
 
 ## Installation
 
 Add this line to your application's Gemfile:
 
 ```ruby
-gem 'structured_logger'
+gem 'logr'
 ```
 
 And then execute:
@@ -18,7 +16,7 @@ And then execute:
 
 Or install it yourself as:
 
-    $ gem install structured_logger
+    $ gem install logr
 
 ## Usage
 
@@ -32,7 +30,7 @@ To install this gem onto your local machine, run `bundle exec rake install`. To 
 
 ## Contributing
 
-Bug reports and pull requests are welcome on GitHub at https://github.com/[USERNAME]/structured_logger.
+Bug reports and pull requests are welcome on GitHub at https://github.com/sspinc/logr.
 
 
 ## License
