+++ lib/logr/metric.rb
@@ -1,4 +1,4 @@
-class StructuredLogger
+module Logr
   Metric = Struct.new(:name, :value, :type) do
     def to_hash; to_h end
   end
