+++ lib/logr/version.rb
@@ -1,3 +1,3 @@
-class StructuredLogger
+module Logr
   VERSION = "0.1.0"
 end
