+++ logr.gemspec
@@ -1,17 +1,17 @@
 # coding: utf-8
 lib = File.expand_path('../lib', __FILE__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
-require 'structured_logger/version'
+require 'logr/version'
 
 Gem::Specification.new do |spec|
-  spec.name          = "structured_logger"
-  spec.version       = StructuredLogger::VERSION
+  spec.name          = "logr"
+  spec.version       = Logr::VERSION
   spec.authors       = ["Peter Marton"]
   spec.email         = ["martonpe@secretsaucepartners.com"]
 
-  spec.summary       = "Structured logging for events and metrics"
-  spec.description   = "Structured logging for events and metrics"
-  spec.homepage      = "https://github.com/sspinc/structured_logger"
+  spec.summary       = "Structured logging with events and metrics"
+  spec.description   = "Structured logging with events and metrics"
+  spec.homepage      = "https://github.com/sspinc/logr"
   spec.license       = "MIT"
 
   spec.files         = `git ls-files -z`.split("\x0").reject { |f| f.match(%r{^(test|spec|features)/}) }
