+++ spec/structured_logger_spec.rb
@@ -1,8 +1,8 @@
 require 'spec_helper'
 
-describe StructuredLogger do
+describe Logr do
   it 'has a version number' do
-    expect(StructuredLogger::VERSION).not_to be nil
+    expect(Logr::VERSION).not_to be nil
   end
 
   it 'does something useful' do
